requirements.txt
# Core ML/CV libraries
ultralytics==8.0.196
opencv-python==4.8.1.78
torch>=2.2.0,<2.9.0
torchvision>=0.17.0,<0.20.0
numpy>=1.25.0,<2.0.0

# Tracking library
deep-sort-realtime==1.3.2

# Utility libraries
imutils==0.5.4
pillow==10.0.0

# Logging and configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Video processing (optional)
imageio==2.31.1
imageio-ffmpeg==0.4.8

# Development and testing
pytest==7.4.0
black==23.7.0
flake8==6.0.0

# System utilities
psutil==5.9.5
tqdm==4.65.0